Bug Fix Issue #95 Assert.Response hangs when assertion fails
This issue occurs based on the assertion library taking over and throwing an exception when it's base assertions fail, this unfortunately means as soon as a failure occurs no further code in expresso in that assert.response.   I've added try/catch blocks to catch the error assert throws and then call the check() method before issuing an assert.ok(false, err.message) which will of course cause the exception to then be re-thrown.